apps/flask_apps/purchase-requisition/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg[binary]>=3.1
gunicorn==22.0.0
pandas==2.2.2
openpyxl==3.1.5
prometheus-flask-exporter==0.23.1
python-dotenv==1.0.1
PyYAML==6.0.1
resend>=2.0.0
